Add changes required for arm64 vs tools to work
First batch of changes to permit VS2017 ARM64 tools to build required elements.

+++ ArmPkg/Drivers/ArmGic/ArmGicLib.c
@@ -148,10 +148,8 @@ ArmGicSendSgiTo (
 {
   MmioWrite32 (
     GicDistributorBase + ARM_GIC_ICDSGIR,
-    ((TargetListFilter & 0x3) << 24) |
-    ((CPUTargetList & 0xFF) << 16)   |
-    (SgiId & 0xF)
-    );
+    (UINT32)(((TargetListFilter & 0x3) << 24) | ((CPUTargetList & 0xFF) << 16) | (SgiId & 0xF))
+    );      // MU_CHANGE
 }
 
 /*
