+++ ArmPkg/Drivers/CpuDxe/CpuDxe.c
@@ -226,7 +226,7 @@ InitializeDma (
   IN OUT  EFI_CPU_ARCH_PROTOCOL  *CpuArchProtocol
   )
 {
-  CpuArchProtocol->DmaBufferAlignment = ArmCacheWritebackGranule ();
+  CpuArchProtocol->DmaBufferAlignment = (UINT32)ArmCacheWritebackGranule ();  // MU_CHANGE - ARM64 VS change
 }
 
 /**
